[#121] feat: 파일이름변경 및 추가 메서드 생성
