+++ rest/src/main/java/com/wherecar/rest/domain/Car.java
@@ -38,7 +38,7 @@ public class Car extends BaseEntity{
     @JoinColumn(name="company_id")
     private Company company;
 
-    @OneToOne(fetch = FetchType.LAZY)
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name="geo_info_id")
     private GeoInfo geoInfo;
