+++ rest/src/main/java/com/wherecar/rest/geo/domain/GeoInfo.java
@@ -1,6 +1,7 @@
 package com.wherecar.rest.geo.domain;
 
 import com.wherecar.rest.domain.BaseEntity;
+import com.wherecar.rest.domain.Company;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -19,18 +20,24 @@ public class GeoInfo extends BaseEntity {
     @Column(name="geo_info_id")
     private Long id;
 
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name="company_id")
+    private Company company;
+
+    @Column(name = "name")
+    private String name;
+
     @Column(name = "geo_event_type")
     private String geoEventType;
 
     @Column(name = "geo_range")
     private String geoRange;
 
-
     @Column(name = "latitude")
-    private Integer latitude;
+    private Double latitude;
 
     @Column(name = "longitude")
-    private Integer longitude;
+    private Double longitude;
 
     @Column(name = "on_time")
     private LocalDateTime onTime;
@@ -42,15 +49,19 @@ public class GeoInfo extends BaseEntity {
         this.geoEventType = geoEventType;
     }
 
+    public void changeName(String name) {
+        this.name = name;
+    }
+
     public void changeGeoRange(String geoRange) {
         this.geoRange = geoRange;
     }
 
-    public void changeLatitude(Integer latitude) {
+    public void changeLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public void changeLongitude(Integer longitude) {
+    public void changeLongitude(Double longitude) {
         this.longitude = longitude;
     }
 }
