+++ rest/src/main/java/com/wherecar/rest/geo/domain/GeoLog.java
@@ -19,8 +19,8 @@ public class GeoLog extends BaseEntity {
     @Column(name="geo_log_id")
     private Long id;
 
+    @Column(name = "mdn")
     private String mdn;
-
 
     @Column(name = "o_time")
     private LocalDateTime oTime;
@@ -29,10 +29,10 @@ public class GeoLog extends BaseEntity {
     private String gpsCondition;
 
     @Column(name = "latitude")
-    private Integer latitude;
+    private Double latitude;
 
     @Column(name = "longitude")
-    private Integer longitude;
+    private Double longitude;
 
     @Column(name = "angle")
     private Integer angle;
@@ -62,11 +62,11 @@ public class GeoLog extends BaseEntity {
         this.gpsCondition = gpsCondition;
     }
 
-    public void changeLatitude(Integer latitude) {
+    public void changeLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public void changeLongitude(Integer longitude) {
+    public void changeLongitude(Double longitude) {
         this.longitude = longitude;
     }
