+++ rest/src/main/java/com/wherecar/rest/geo/dto/emulator/ControlResponse.java
@@ -1,4 +1,4 @@
-package com.wherecar.rest.dto;
+package com.wherecar.rest.geo.dto.emulator;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class ControlGeoList {
+public class ControlResponse {
     private String ctrId;  // 제어 아이디
     private String ctrCd;  // 제어 코드
     private String ctrVal; // 제어 값
